refactor: donation status/redemption extracted to it's own component
