+++ src/App.js
@@ -34,7 +34,6 @@ export default class App extends Component {
 
     const { venues } = this.state;
       return (<div>
-        <VenueProfile />
         { this.state.signup ?
           < Signup /> :
           <div className="App">
