+++ src/components/venueProfile.js
@@ -1,4 +1,5 @@
 import React, {Component} from 'react';
+import DonationStatus from './donationStatus';
 
 const venue1 ={
   id: 1,
@@ -31,46 +32,15 @@ export default class VenueProfile extends Component {
     })
   }
 
-  // https://enigmatic-badlands-83570.herokuapp.com/api/v1/venues/${venueId}/donations
-
-  redeemDonation = id => {
-
-    const body = JSON.stringify({ redeemed: true })
-
-    const venueId = this.state.venue.id;
-    const donationId = id;
-
-    fetch(`http://localhost:3000/api/v1/venues/${venueId}/donations/${donationId}`, {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: body
-    })
-    .then((response) => {
-      return response.json()
-    })
-
-  }
-
   render() {
     const { venue } = this.state;
     const { donations } = this.state;
     const openDonations = donations.filter(donation => donation.redeemed !== true)
 
-    const list = openDonations.map((donation) => {
-      return(
-        <div className="donationList">
-          <p key={donation.id}>
-            <li>{donation.amount} {donation.passphrase}</li>
-            <button onClick={()=> {this.redeemDonation(donation.id)}}>Redeem</button>
-          </p>
-        </div>
-      )
-    })
+    const list = openDonations.map(donation => <DonationStatus donation={donation} venue={venue} />)
 
     return(
-      <div>
+      <div className="donationList">
         <h1>{venue.name}</h1>
         <ul>{list}</ul>
       </div>
